assets/prediction_assets/requirements.txt.txt
numpy==1.20.1
pandas==1.2.4
pybaseball==2.2.1
requests==2.25.1
scikit_learn==1.1.2
tqdm==4.59.0
MLB-StatsAPI
pybaseball

